refactor(api): package chip services
