+++ src/qdash/api/services/chip/__init__.py
@@ -1,0 +1,6 @@
+"""Chip-related API services."""
+
+from qdash.api.services.chip.initializer import ChipInitializer
+from qdash.api.services.chip.service import ChipService, get_task_names
+
+__all__ = ["ChipInitializer", "ChipService", "get_task_names"]
